socialnetwork/requirements.txt
automium==0.2.6
automium-web==0.1.1
Click==7.0
Django==2.2.6
django-rest-auth==0.9.5
djangorestframework==3.10.3
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
MarkupSafe==1.1.1
pkg-resources==0.0.0
post==2019.4.13
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
views==0.3
Werkzeug==0.16.0
